Extract Response from net_traits/lib.rs
