+++ components/net/fetch/request.rs
@@ -15,8 +15,8 @@ use hyper::header::{QualityItem, q, qitem, Referer as RefererHeader, UserAgent};
 use hyper::method::Method;
 use hyper::mime::{Attr, Mime, SubLevel, TopLevel, Value};
 use hyper::status::StatusCode;
-use net_traits::{AsyncFetchListener, CacheState, HttpsState, Response};
-use net_traits::{ResponseType, Metadata, TerminationReason};
+use net_traits::response::{CacheState, HttpsState, Response, ResponseType, TerminationReason};
+use net_traits::{AsyncFetchListener, Metadata};
 use resource_task::CancellationListener;
 use std::ascii::AsciiExt;
 use std::cell::RefCell;
