+++ components/net_traits/lib.rs
@@ -28,12 +28,9 @@ use hyper::header::{ContentType, Headers};
 use hyper::http::RawStatus;
 use hyper::method::Method;
 use hyper::mime::{Attr, Mime};
-use hyper::status::StatusCode;
 use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
 use msg::constellation_msg::{PipelineId};
 use serde::{Deserializer, Serializer};
-use std::cell::RefCell;
-use std::rc::Rc;
 use std::thread;
 use url::Url;
 use util::mem::HeapSizeOf;
@@ -42,100 +39,8 @@ use websocket::header;
 pub mod hosts;
 pub mod image_cache_task;
 pub mod net_error_list;
+pub mod response;
 pub mod storage_task;
-
-/// [Response type](https://fetch.spec.whatwg.org/#concept-response-type)
-#[derive(Clone, PartialEq, Copy)]
-pub enum ResponseType {
-    Basic,
-    CORS,
-    Default,
-    Error,
-    Opaque,
-    OpaqueRedirect
-}
-
-/// [Response termination reason](https://fetch.spec.whatwg.org/#concept-response-termination-reason)
-#[derive(Clone, Copy)]
-pub enum TerminationReason {
-    EndUserAbort,
-    Fatal,
-    Timeout
-}
-
-/// The response body can still be pushed to after fetch
-/// This provides a way to store unfinished response bodies
-#[derive(Clone)]
-pub enum ResponseBody {
-    Empty, // XXXManishearth is this necessary, or is Done(vec![]) enough?
-    Receiving(Vec<u8>),
-    Done(Vec<u8>),
-}
-
-/// [Cache state](https://fetch.spec.whatwg.org/#concept-response-cache-state)
-#[derive(Clone)]
-pub enum CacheState {
-    None,
-    Local,
-    Validated,
-    Partial
-}
-
-/// [Https state](https://fetch.spec.whatwg.org/#concept-response-https-state)
-#[derive(Clone)]
-pub enum HttpsState {
-    None,
-    Deprecated,
-    Modern
-}
-
-pub enum ResponseMsg {
-    Chunk(Vec<u8>),
-    Finished,
-    Errored
-}
-
-/// A [Response](https://fetch.spec.whatwg.org/#concept-response) as defined by the Fetch spec
-#[derive(Clone)]
-pub struct Response {
-    pub response_type: ResponseType,
-    pub termination_reason: Option<TerminationReason>,
-    pub url: Option<Url>,
-    pub url_list: Vec<Url>,
-    /// `None` can be considered a StatusCode of `0`.
-    pub status: Option<StatusCode>,
-    pub headers: Headers,
-    pub body: ResponseBody,
-    pub cache_state: CacheState,
-    pub https_state: HttpsState,
-    /// [Internal response](https://fetch.spec.whatwg.org/#concept-internal-response), only used if the Response
-    /// is a filtered response
-    pub internal_response: Option<Rc<RefCell<Response>>>,
-}
-
-impl Response {
-    pub fn network_error() -> Response {
-        Response {
-            response_type: ResponseType::Error,
-            termination_reason: None,
-            url: None,
-            url_list: vec![],
-            status: None,
-            headers: Headers::new(),
-            body: ResponseBody::Empty,
-            cache_state: CacheState::None,
-            https_state: HttpsState::None,
-            internal_response: None
-        }
-    }
-
-    pub fn is_network_error(&self) -> bool {
-        match self.response_type {
-            ResponseType::Error => true,
-            _ => false
-        }
-    }
-}
 
 /// Image handling.
 ///
@@ -197,7 +102,7 @@ impl LoadData {
 
 /// Interface for observing the final response for an asynchronous fetch operation.
 pub trait AsyncFetchListener {
-    fn response_available(&self, response: Response);
+    fn response_available(&self, response: response::Response);
 }
 
 /// A listener for asynchronous network events. Cancelling the underlying request is unsupported.
